reactor: refactoring permissiong keys to api package
