+++ core/api/options.go
@@ -24,7 +24,6 @@
 package api
 
 import (
-	"infini.sh/framework/core/security"
 	"infini.sh/framework/core/util"
 )
 
@@ -34,7 +33,7 @@ type Option func(*HandlerOptions)
 // Define HandlerOptions to hold the state of all options
 type HandlerOptions struct {
 	RequireLogin      bool
-	RequirePermission []security.PermissionKey
+	RequirePermission []PermissionKey
 	OptionLogin       bool
 	Resource          string
 	Action            string
@@ -78,12 +77,12 @@ func RequireLogin() Option {
 	}
 }
 
-func RequirePermission(permissions ...security.PermissionKey) Option {
+func RequirePermission(permissions ...PermissionKey) Option {
 	return func(o *HandlerOptions) {
 		o.RequireLogin = true
 		o.OptionLogin = false
 		if o.RequirePermission == nil {
-			o.RequirePermission = []security.PermissionKey{}
+			o.RequirePermission = []PermissionKey{}
 		}
 
 		for _, v := range permissions {
